# contracts/http/openapi.yaml
openapi: 3.1.0
info:
  title: Synapse - API HTTP
  version: 1.0.0
  summary: Contrato REST e SSE entre o frontend e a api do Synapse.
  description: |
    Fonte única do contrato HTTP da Sprint 1. Nenhuma rota da Sprint 1 existe fora
    deste arquivo: `frontend` e `api` são escritos em paralelo contra ele e só se
    encontram na integração.

    Sete operações, todas sobre o recurso `job`, que é a unidade de trabalho do
    sistema e o que a `api` possui de ponta a ponta. O ciclo é assíncrono: a criação
    devolve o job em processamento, o acompanhamento é por stream de eventos e o
    resultado aparece na consulta do job quando existe.

    **Fluxo de trabalho padrão:** As sete operações formam um ciclo, não uma lista solta de
    endpoints. Da submissão ao fechamento:

    * Job é a entidade que envelopa todo o fluxo de trabalho, desde a validação da regra até o resultado da simulação.

    1. `POST /jobs` submete a regra e cria o job. Na Sprint 1 a regra já nasce
       validada e estruturada a partir do formulário, então o job entra direto em
       `gerando_regra`: submissão de formulário não passa por uma etapa
       intermediária de confirmação do usuário.
    2. O cliente abre conexão SSE `GET /jobs/{id}/events` logo em seguida. O primeiro
       evento é a fotografia do estado atual; os demais chegam conforme o job avança,
       inclusive o que os próximos passos disparam. A conexão pode cair e ser
...
       reabrir devolve a mesma fotografia de novo.
...

...

...

...

...
servers:
...
security:
  - keycloakOidc: []
...
paths:
...
  /jobs/{id}/parameters:
    parameters:
      - $ref: "#/components/parameters/JobId"
    post:
      tags: [jobs]
      operationId: confirmarParametros
      summary: Confirma ou corrige a representação da regra e retoma o processamento.
      description: |
        O que o usuário confirma na tela de validação é a representação da regra, e
        nunca código. A `api` grava uma versão nova de regra -
        correção nunca altera a versão anterior - e publica `parametros-confirmados`,
        que retoma o grafo do codegen no nó de confirmação.

        Aceita também `orcamento` e `competencias`, porque a mesma tela permite
        ajustá-los antes de simular, inclusive no reprocessamento de uma regra
        arquivada. Ausentes, os valores atuais do job permanecem.

        Recusa por completude e por coerência, com códigos distintos: núcleo
        incompleto é campo faltando; regra incoerente é a regra ter todos os campos e
        ainda assim não descrever nada simulável, como uma faixa cujo limite inferior
        é maior que o superior. A segunda não rende sugestão de adaptação, e sim
        revisão manual.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmarParametrosRequisicao"
      responses:
        "202":
          description: |
            Parâmetros confirmados e processamento retomado. O corpo é o job já em
            `gerando_regra`, com `regra` na versão que acabou de ser confirmada:
            correção do usuário vira versão nova, e é o `id` dela que identifica o que
            será simulado.

            **`simulacao`, `finalizado_em` e `motivo` continuam ausentes.** A operação
            só é aceita com o job em `aguardando_confirmacao_parametros`, estado em que
            nenhuma simulação concluiu. A geração de código e a execução no sandbox são
            assíncronas e começam agora; o desfecho chega pelo stream e, depois dele,
            por `GET /jobs/{id}`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
              examples:
                parametrosConfirmados:
                  summary: Parâmetros confirmados com correção, geração em curso
                  value:
                    id: "3f2b1c40-0d18-4a51-9f2e-6c1d9a77b021"
                    status: gerando_regra
                    origem: formulario
                    competencias: ["2025-11"]
                    orcamento: 485000.0
                    criado_em: "2025-11-24T14:02:00Z"
                    iniciado_em: "2025-11-24T14:03:10Z"
                    submissao_id: "b81e0f4c-52a9-4f0b-8a3d-7c2e5d10ab93"
                    regra:
                      id: "1f2e3d4c-5b6a-4978-8d7c-6e5f4a3b2c1d"
                      versao: 2
                      origem: confirmacao_usuario
                      representacao:
                        nucleo:
                          vigencia: { inicio: "2025-11", fim: "2025-11" }
                          loja: ["13"]
                          marca: ["10", "20"]
                          cargo: ["100", "300"]
                          percentual: 0.03
                        especificacoes: []
                      criada_em: "2025-11-24T14:03:10Z"
        "400":
          $ref: "#/components/responses/RequisicaoInvalida"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/SemPermissao"
        "404":
          $ref: "#/components/responses/JobNaoEncontrado"
        "409":
          description: |
            O job não está aguardando confirmação de parâmetros. Confirmar duas vezes
            ou confirmar um job já finalizado cai aqui.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Erro"
              examples:
                estadoInvalido:
                  value:
                    codigo: estado_invalido
                    mensagem: Este job não está aguardando confirmação de parâmetros.
        "422":
          description: |
            A regra não pode ser simulada como está. `codigo` distingue campo ausente
            de conflito entre elementos, e `elementos` localiza o problema na parte da
            regra que o causou.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Erro"
              examples:
                nucleoIncompleto:
                  summary: Campo obrigatório ausente
                  value:
                    codigo: nucleo_incompleto
                    mensagem: A regra precisa do percentual de comissionamento para ser simulada.
                    elementos:
                      - ref: nucleo.percentual
                        motivo: Campo obrigatório não informado.
                regraIncoerente:
                  summary: Elementos que se contradizem
                  value:
                    codigo: regra_incoerente
                    mensagem: A vigência termina antes de começar.
                    elementos:
                      - ref: nucleo.vigencia
                        motivo: A competência final é anterior à inicial.
...
components:
  securitySchemes:
    keycloakOidc:
      type: openIdConnect
      openIdConnectUrl: http://localhost:8081/realms/synapse/.well-known/openid-configuration
      description: |
        Autenticação OIDC emitida pelo Keycloak. O cliente envia o access token como
        `Authorization: Bearer <token>`; a API valida emissor, assinatura e expiração.
        A conta local é associada ao claim `sub`, e o acesso a um job confere posse,
        não apenas autenticação. O papel `auditor` no realm autoriza consulta além da
        própria conta.
  parameters:
    JobId:
      name: id
      in: path
      required: true
      description: Identificador do job.
      schema:
        type: string
        format: uuid
        examples: ["3f2b1c40-0d18-4a51-9f2e-6c1d9a77b021"]
  responses:
    RequisicaoInvalida:
      description: O corpo ou os parâmetros da requisição não têm a forma esperada.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Erro"
          examples:
            requisicaoInvalida:
              value:
                codigo: requisicao_invalida
                mensagem: O campo orcamento precisa ser um número.
    NaoAutenticado:
      description: Sessão ausente, expirada ou inválida.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Erro"
          examples:
            naoAutenticado:
              value:
                codigo: nao_autenticado
                mensagem: Sessão ausente ou expirada.
    SemPermissao:
      description: |
        A sessão é válida, mas o solicitante não é dono do job nem tem papel que
        autorize a operação sobre ele.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Erro"
          examples:
            semPermissao:
              value:
                codigo: sem_permissao
                mensagem: Você não tem acesso a este job.
    JobNaoEncontrado:
      description: Não existe job com o identificador informado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Erro"
          examples:
            jobNaoEncontrado:
              value:
                codigo: job_nao_encontrado
                mensagem: Job não encontrado.
  schemas:
    Erro:
      type: object
      title: Erro
      description: |
        Corpo de erro de toda operação. `codigo` é o que o cliente decide em cima;
        `mensagem` é texto que o domínio decidiu expor e nunca detalhe interno, como
        stack trace, mensagem de driver ou nome de tabela.

        `elementos` existe porque interrupção sem motivo é o pior caso para o usuário
        leigo: todo caminho de parada devolve razão localizada na parte da regra que a
        causou, em vez de "não foi possível processar".
      required: [codigo, mensagem]
      properties:
        codigo:
          $ref: "#/components/schemas/CodigoErro"
        mensagem:
          type: string
          minLength: 1
          description: Explicação em linguagem do usuário, própria para exibição direta.
        elementos:
          type: array
          description: |
            Partes da regra que motivaram a recusa. Ausente quando o erro não é sobre
            a regra. Um conflito entre dois elementos rende uma entrada por elemento
            envolvido.
          items:
            type: object
            required: [ref, motivo]
            properties:
              ref:
                $ref: "../domain/comum.schema.json#/$defs/elemento_ref"
              motivo:
                type: string
                minLength: 1
                description: O que falta ou o que conflita neste elemento.
    CodigoErro:
      type: string
      title: Código de erro
      description: |
        Vocabulário de recusa. Cada código tem um status HTTP próprio, e dois motivos
        diferentes nunca compartilham código: `nucleo_incompleto` é dado da regra
        faltando na entrada, `simulacao_inviavel` é regra completa e coerente cujo
        custo não cabe no orçamento. Vocabulário aberto a valores novos, sempre de
        forma aditiva.
      enum:
        - requisicao_invalida
        - nao_autenticado
        - sem_permissao
        - job_nao_encontrado
        - estado_invalido
        - simulacao_inviavel
        - nucleo_incompleto
        - regra_incoerente
    StatusJob:
      type: string
      title: Estado do job
      description: |
        Estados da máquina de estados do job, de que a `api` é dona. `simulacao_inviavel` e
        `erro` são coisas distintas: o primeiro é uma simulação que terminou bem e
        produziu um número que não cabe no orçamento; o segundo é o job que não chegou
        a produzir número.
      enum:
        - aguardando_transcricao
        - aguardando_confirmacao_parametros
        - gerando_regra
        - simulando
        - simulacao_inviavel
        - aguardando_decisao_usuario
        - liberado
        - cancelado
        - arquivado
        - erro
    OrigemJob:
      type: string
      title: Origem do job
      description: |
        Como o job nasceu, no mesmo vocabulário de `contracts/events/regra-submetida`.
        É o que determina o ponto de entrada do grafo. Na Sprint 1 apenas `formulario`
        é aceito em `POST /jobs`; `reprocessamento` é produzido por
        `POST /jobs/{id}/reprocessar` e `voz`.
      enum: [formulario, voz, reprocessamento]
...
    Veredito:
      type: string
      title: Veredito de viabilidade
      description: |
        Viabilidade orçamentária, apurada pelo worker fora do container a partir do
        total simulado e do orçamento aplicado. Ausente quando a execução não produziu
        número que o sustente.
      enum: [viavel, inviavel, indeterminado]
    StatusResultado:
      type: string
      title: Desfecho da execução
      description: |
        Mesmo vocabulário de `resultados_simulacao.status`. `assercao_violada` é o
        código ter rodado e produzido números que uma invariante reprovou: é erro do
        código gerado e invalida o número, diferente de `veredito: inviavel`.
      enum: [sucesso, assercao_violada, erro_codigo, erro_infra]
    Competencias:
      type: array
      title: Competências do job
      description: |
        Competências históricas a simular. O job cobre o período inteiro numa única
        simulação, com os meses agregados; não existe uma simulação por competência. A
        contribuição de cada mês sai na quebra `competencia` da decomposição. Não
        precisa ser contíguo. Nunca vazio: um job que não simula nada não é job.
      minItems: 1
      items:
        $ref: "../domain/comum.schema.json#/$defs/competencia"
      examples:
        - ["2025-08", "2025-11"]
...
    ConfirmarParametrosRequisicao:
      type: object
      title: Confirmação dos parâmetros
      description: |
        Corpo de `POST /jobs/{id}/parameters`. `regra` é a representação confirmada ou
        corrigida pelo usuário, no mesmo formato que o banco e o código gerado usam.
        Correção gera uma versão nova de regra; a anterior permanece.
      required: [regra]
      properties:
        regra:
          $ref: "../domain/representacao-regra.schema.json"
          description: |
            Núcleo e especificações juntos. Na Sprint 1 o formulário só permite
            informar o núcleo, então `especificacoes` chega como array vazio; a
            limitação está no que pode ser inserido, nunca no que é simulado daquilo
            que foi inserido.
        orcamento:
          $ref: "../domain/comum.schema.json#/$defs/brl"
          description: |
            Orçamento ajustado na tela de confirmação. Ausente mantém o orçamento
            atual do job.
        competencias:
          $ref: "#/components/schemas/Competencias"
          description: |
            Competências ajustadas na tela de confirmação. Ausente mantém as
            competências atuais do job.
      examples:
        - regra:
            nucleo:
              vigencia: { inicio: "2025-11", fim: "2025-11" }
              loja: ["13"]
              marca: ["10", "20"]
              cargo: ["100", "300"]
              percentual: 0.025
            especificacoes: []
          orcamento: 485000.0
...
    Job:
      type: object
      title: Job
      description: |
        Estado corrente do job e os artefatos que ele já produziu. Os campos opcionais
        aparecem conforme o job avança; a ausência de `simulacao` significa que ainda
        não há resultado, e não que o resultado é vazio.
      required: [id, status, origem, competencias, orcamento, criado_em]
      properties:
        id:
          type: string
          format: uuid
          description: Identificador do job, e a chave de correlação em toda a observabilidade.
        status:
          $ref: "#/components/schemas/StatusJob"
        origem:
          $ref: "#/components/schemas/OrigemJob"
        competencias:
          $ref: "#/components/schemas/Competencias"
        orcamento:
          $ref: "../domain/comum.schema.json#/$defs/brl"
          description: Orçamento do período aplicado a este job.
        criado_em:
          type: string
          format: date-time
        iniciado_em:
          type: string
          format: date-time
          description: Ausente até o processamento começar.
        finalizado_em:
          type: string
          format: date-time
          description: Ausente até o job atingir estado terminal.
        job_origem_id:
          type: string
          format: uuid
          description: |
            Job de que este nasceu por reprocessamento. Ausente no job original.
            Exatamente uma das duas procedências existe: `submissao_id` quando o job
            nasceu de gente, `job_origem_id` quando nasceu de outro job.
        submissao_id:
          type: string
          format: uuid
          description: Submissão que originou o job. Ausente quando a origem é reprocessamento.
        motivo:
          type: string
          description: |
            Razão localizada da parada, quando o job parou. Existe porque interrupção
            sem motivo é inaceitável para o usuário leigo: regra fora do domínio,
            elemento sem implementação correspondente, asserção violada, campo
            obrigatório ausente e conflito entre elementos são casos distintos e
            chegam aqui distinguidos. Ausente fora de paradas.
        regra:
          $ref: "#/components/schemas/RegraVersionada"
          description: Versão corrente da regra. Ausente enquanto nenhuma versão foi formada.
        simulacao:
          $ref: "#/components/schemas/Simulacao"
          description: Simulação corrente do job. Ausente enquanto a execução não terminou.
...
    RegraVersionada:
      type: object
      title: Versão da regra
      description: |
        Uma versão da regra dentro do job. Regra é imutável depois de usada: correção
        do usuário e sugestão de adaptação criam versões novas, e é o `id` daqui que
        identifica univocamente o que foi simulado.
      required: [id, versao, origem, representacao, criada_em]
      properties:
        id:
          type: string
          format: uuid
        versao:
          type: integer
          minimum: 1
          description: Sequencial dentro do job.
        origem:
          type: string
          enum: [confirmacao_usuario, sugestao_adaptacao, reprocessamento]
          description: De onde esta versão veio.
        representacao:
          $ref: "../domain/representacao-regra.schema.json"
        criada_em:
          type: string
          format: date-time
    Simulacao:
      type: object
      title: Simulação
      description: |
        A simulação do job e, quando houve sucesso, o resultado que o sandbox produziu.
        `status` e `veredito` respondem perguntas diferentes: o primeiro diz se o
        número vale, o segundo diz se ele cabe no orçamento.
      required: [id, criado_em, status, flag_baixa_rastreabilidade]
      properties:
        id:
          type: string
          format: uuid
          description: Id da simulação exigido pela trilha de auditoria.
        regra_id:
          type: string
          format: uuid
          description: Versão da regra que produziu esta simulação. Opcional para compatibilidade com respostas anteriores.
        criado_em:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/StatusResultado"
        veredito:
          $ref: "#/components/schemas/Veredito"
          description: Ausente quando o status não é sucesso.
        flag_baixa_rastreabilidade:
          type: boolean
          description: |
            Verdadeiro quando o modelo não devolveu justificativa utilizável ou quando
            a narrativa reprovou na conferência numérica. Falha de
            explicação degrada a rastreabilidade e não invalida o número; falha de
            cobertura ou de asserção invalida.
        resultado:
          $ref: "../domain/resultado-simulacao.schema.json"
          description: |
            Totais, desfecho das asserções e decomposição da diferença. Ausente quando
            o status não é sucesso, porque aí não há número. É a fonte dos valores
            exibidos: o frontend apresenta, não calcula.
...